legacy-python/requirements.txt
pywebview>=5.0
pystray>=0.19
Pillow>=10.0
requests>=2.31
pywin32>=306
